# medical-dictation-backend/docs/api/report-service-openapi.json
{
  "components": {
    "schemas": {
...
      "CreateReportRequest": {
        "additionalProperties": false,
        "properties": {
          "co_author_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Co Author Ids",
            "type": "array"
          },
          "content": {
            "$ref": "#/components/schemas/ReportContent"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "source_session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Session Id"
          }
        },
        "required": [
          "content",
          "patient_id"
        ],
        "title": "CreateReportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Icd10Code": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 8,
            "minLength": 3,
            "pattern": "^[A-Z][0-9]{2}(\\.[0-9A-Z]{1,4})?$",
            "title": "Code",
            "type": "string"
          },
          "display": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display"
          }
        },
        "required": [
          "code"
        ],
        "title": "Icd10Code",
        "type": "object"
      },
...
      "LocalizedText": {
        "additionalProperties": false,
        "properties": {
          "en": {
            "title": "En",
            "type": "string"
          },
          "uk": {
            "title": "Uk",
            "type": "string"
          }
        },
        "required": [
          "uk",
          "en"
        ],
        "title": "LocalizedText",
        "type": "object"
      },
...
      "ReadPurpose": {
        "enum": [
          "clinical_continuity",
          "audit",
          "legal",
          "qa_review",
          "consultation"
        ],
        "title": "ReadPurpose",
        "type": "string"
      },
...
      "ReportContent": {
        "additionalProperties": false,
        "description": "The full content_jsonb body for one version.",
        "properties": {
          "encounter_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Date"
          },
          "icd10_codes": {
            "items": {
              "$ref": "#/components/schemas/Icd10Code"
            },
            "title": "Icd10 Codes",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "template_schema_version": {
            "minimum": 1.0,
            "title": "Template Schema Version",
            "type": "integer"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "template_id",
          "template_schema_version"
        ],
        "title": "ReportContent",
        "type": "object"
      },
      "ReportCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "version_id",
          "version_number",
          "status"
        ],
        "title": "ReportCreatedResponse",
        "type": "object"
      },
      "ReportEnvelope": {
        "additionalProperties": false,
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "co_author_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Co Author Ids",
            "type": "array"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportContent"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_version_id": {
            "format": "uuid",
            "title": "Current Version Id",
            "type": "string"
          },
          "current_version_number": {
            "title": "Current Version Number",
            "type": "integer"
          },
          "encounter_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Date"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "icd10_codes": {
            "items": {
              "type": "string"
            },
            "title": "Icd10 Codes",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patient_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "patient_name_redacted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name Redacted"
          },
          "primary_author_id": {
            "format": "uuid",
            "title": "Primary Author Id",
            "type": "string"
          },
          "section_labels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SectionLabel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Labels"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "current_version_id",
          "current_version_number",
          "primary_author_id",
          "co_author_ids",
          "patient_id",
          "patient_name_redacted",
          "title",
          "icd10_codes",
          "encounter_date",
          "created_at",
          "updated_at",
          "finalized_at",
          "signed_at",
          "cancelled_at"
        ],
        "title": "ReportEnvelope",
        "type": "object"
      },
      "ReportSection": {
        "additionalProperties": false,
        "properties": {
          "field_specific_metadata": {
            "additionalProperties": true,
            "title": "Field Specific Metadata",
            "type": "object"
          },
          "icd10": {
            "items": {
              "$ref": "#/components/schemas/Icd10Code"
            },
            "title": "Icd10",
            "type": "array"
          },
          "section_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Section Key",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "transcript_segment_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Transcript Segment Ids",
            "type": "array"
          }
        },
        "required": [
          "section_key"
        ],
        "title": "ReportSection",
        "type": "object"
      },
...
      "SectionLabel": {
        "additionalProperties": false,
        "description": "Human-readable, localized title for one report section.\n\nTemplates are per-language (a template row is either ``uk`` or ``en``),\nso the section ``name`` is a single string in the template's own\nlanguage. We mirror it into BOTH ``uk`` and ``en`` here \u2014 the same\nbehaviour the frontend's ``toStudioTemplate`` uses \u2014 so the SPA/PDF can\nrender section titles without re-fetching the template, and historical\nreports carry their labels even if the template later changes.",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "section_key": {
            "title": "Section Key",
            "type": "string"
          }
        },
        "required": [
          "section_key",
          "name"
        ],
        "title": "SectionLabel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Sprint-06 templates + sprint-08 reports / versions / diff / search.",
    "title": "Report Service",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reports": {
      "post": {
        "operationId": "create_report_v1_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "`patient_not_found`, or sprint-13 field-metadata validation: `field_metadata_invalid` / `choice_value_unknown` (section-addressed problems in `problems[]`)."
          }
        },
        "summary": "Create Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/v1/reports/{report_id}": {
      "get": {
        "operationId": "get_report_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "description": "Required for non-author reads.",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReadPurpose"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required for non-author reads.",
              "title": "Purpose"
            }
          },
          {
            "in": "query",
            "name": "include_content",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Content",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/v1/reports/{report_id}/pdf": {
      "get": {
        "operationId": "get_report_pdf_v1_reports__report_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "description": "Required for non-author reads.",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReadPurpose"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required for non-author reads.",
              "title": "Purpose"
            }
          },
          {
            "description": "'clean' is only honoured for signed reports; non-signed reports are always draft.",
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "draft",
              "description": "'clean' is only honoured for signed reports; non-signed reports are always draft.",
              "enum": [
                "draft",
                "clean"
              ],
              "title": "Variant",
              "type": "string"
            }
          },
          {
            "description": "Render language; falls back to the report language, else 'uk'.",
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "uk",
                    "en"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Render language; falls back to the report language, else 'uk'.",
              "title": "Lang"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render the current version as a PDF (draft watermark for non-signed reports; 409 only for cancelled).",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}